+++ src/app/api/checklists/[checkListId]/archive/route.ts
@@ -3,26 +3,18 @@ import { NextRequest, NextResponse } from "next/server";
 
 const prisma = new PrismaClient();
 
-// アーカイブされたチェックリストを復元する
+// チェックリストのアーカイブ
 export const POST = async (
   req: NextRequest,
-  { params }: { params: { id: string } }
+  { params }: { params: { checkListId: string } }
 ) => {
-  /**
-   * アーカイブされたチェックリストを検索
-   * where: { archivedAt: { not: null } }
-   * アクティブなチェックリストを検索
-   * where: { archivedAt: null }
-   *
-   * 復元する時は、archivedAtをnullに設定する
-   */
   try {
     const checkList = await prisma.checkLists.update({
       where: {
-        id: parseInt(params.id),
+        id: parseInt(params.checkListId),
       },
       data: {
-        archivedAt: null,
+        archivedAt: new Date(),
       }
     })
 
@@ -30,7 +22,7 @@ export const POST = async (
   } catch (error) {
     console.error('Error', error)
     return NextResponse.json(
-      { error: 'Error restoring checklist' },
+      { error: 'Error archiving checklist' },
       { status: 500 }
     )
   }
